Finished the book

+++ src/utils.h
@@ -11,6 +11,8 @@ inline double degrees_to_radius(double degrees) { return degrees * pi / 180; }
 
 inline double random_double() { return std::rand() / (RAND_MAX + 1.0); }
 
-inline double random_double(int min, int max) { return min + (max - min) * random_double(); }
+inline double random_double(double min, double max) {
+    return min + (max - min) * random_double();
+}
 
 #endif
